Add: Compress Night duty ui, adjusted backend return structure for compress grid endpoint
